requirements.txt
appdirs==1.4.4
APScheduler==3.6.3
attrs==20.3.0
autobahn==21.3.1
Automat==20.2.0
autopep8==1.5.7
beautifulsoup4==4.9.3
billiard==3.6.3.0
bs4==0.0.1
cachetools==4.2.1
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
coala-utils==0.5.1
constantly==15.1.0
cryptography==3.4.7
cssselect==1.1.0
cssutils==2.2.0
dateparser==1.0.0
fake-useragent==0.1.11
future==0.18.2
hyperlink==21.0.0
idna==2.10
importlib-metadata==3.7.3
incremental==21.3.0
keyring==23.0.0
knockknock==0.1.8.1
lxml==4.6.2
matrix-client==0.3.2
MechanicalSoup==1.0.0
numpy==1.20.1
oauthlib==3.1.0
pandas==1.2.3
parse==1.19.0
premailer==3.7.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.7.0
pycparser==2.20
pyee==8.1.0
PyJWT==1.7.1
pyOpenSSL==20.0.1
pyppeteer==0.2.5
pyquery==1.4.3
PySocks==1.7.1
python-binance==0.7.11
python-dateutil==2.8.1
python-decouple==3.4
python-telegram-bot==13.4.1
python-twitter==3.5
pytz==2021.1
regex==2021.4.4
requests==2.25.1
requests-html==0.10.0
requests-oauthlib==1.3.0
selenium==3.141.0
service-identity==18.1.0
six==1.15.0
soupsieve==2.2
toml==0.10.2
tornado==6.1
tqdm==4.59.0
tweepy==3.10.0
twilio==6.54.0
Twisted==21.2.0
twitter-scraper==0.4.4
txaio==21.2.1
tzlocal==2.1
ujson==4.0.2
urllib3==1.26.3
w3lib==1.22.0
websocket-client==0.58.0
yagmail==0.14.245
zipp==3.4.1
zope.interface==5.4.0
